requirements.txt
streamlit
google-cloud-bigquery
google-cloud-bigquery-storage
google-auth
pandas
plotly
db-dtypes
pyarrow
